refactor(desktop): organize chat model selector hooks
