+++ desktop/src/hooks/chat/derived/use-chat-launch-catalog.ts
