+++ desktop/src/hooks/plans/workflows/use-plan-handoff-workflow.ts
@@ -8,8 +8,8 @@ import { useSetSessionConfigOptionMutation } from "@anyharness/sdk-react";
 import { PLAN_HANDOFF_DEFAULT_PROMPT } from "@/copy/plans/plan-prompts";
 import { useAgentCatalog } from "@/hooks/agents/derived/use-agent-catalog";
 import { useActiveSessionLaunchState } from "@/hooks/chat/derived/use-active-chat-session-selectors";
-import { useChatLaunchCatalog } from "@/hooks/chat/use-chat-launch-catalog";
-import { useConfiguredLaunchReadiness } from "@/hooks/chat/use-configured-launch-readiness";
+import { useChatLaunchCatalog } from "@/hooks/chat/derived/use-chat-launch-catalog";
+import { useConfiguredLaunchReadiness } from "@/hooks/chat/derived/use-configured-launch-readiness";
 import { useSessionActions } from "@/hooks/sessions/use-session-actions";
 import type { SessionActivationOutcome } from "@/hooks/sessions/session-activation-guard";
 import { isSessionModelAvailabilityInterruption } from "@/hooks/sessions/use-session-model-availability-workflow";
